requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.1
aiohttp==3.9.1
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
aiosqlite==0.19.0
websockets==12.0
gTTS==2.5.1
# TTS==0.13.3  # Replaced by Piper TTS
piper-tts>=1.2.0
